+++ CHANGELOG.md
@@ -6,6 +6,12 @@ and this project adheres to [Semantic Versioning](https://semver.org).
 
 ## [Unreleased](https://github.com/alexmon1989/russia_ddos/compare/2.2.0...HEAD)
 
+### Added
+- `dry run` mode for fast testing purposes
+
+### Fixed
+- Fixed error with keyboard interrupting and threads shutdown process
+
 
 ## [v2.2.0](https://github.com/alexmon1989/russia_ddos/compare/2.1.0...2.2.0)